Localization integrated for painting by story index, painting by manuscript and incipit tools pages.
